fix api-contacts
