+++ controllers/Contacts/addContact.js
@@ -1,9 +1,9 @@
-const Contact = require("../models/Contact");
-const ctrlWrapper = require("../helpers/ctrlWrapper");
+// const Contact = require("../models/Contact");
+// const ctrlWrapper = require("../helpers/ctrlWrapper");
 
-const addContact = async (req, res) => {
-  const result = await Contact.create(req.body);
-  res.status(201).json(result);
-};
+// const addContact = async (req, res) => {
+//   const result = await Contact.create(req.body);
+//   res.status(201).json(result);
+// };
 
-module.exports = { addContact: ctrlWrapper(addContact) };
+// module.exports = { addContact: ctrlWrapper(addContact) };
