+++ controllers/Contacts/getContactById.js
@@ -1,16 +1,16 @@
-const Contact = require("../models/contact");
-const HttpError = require("../helpers/HttpError");
-const ctrlWrapper = require("../helpers/ctrlWrapper");
+// const Contact = require("../models/contact");
+// const HttpError = require("../helpers/HttpError");
+// const ctrlWrapper = require("../helpers/ctrlWrapper");
 
-const getContactById = async (req, res) => {
-  const { contactId } = req.params;
-  const result = await Contact.findById(contactId);
-  if (!result) {
-    throw HttpError(404, "Not found");
-  }
-  res.json(result);
-};
+// const getContactById = async (req, res) => {
+//   const { contactId } = req.params;
+//   const result = await Contact.findById(contactId);
+//   if (!result) {
+//     throw HttpError(404, "Not found");
+//   }
+//   res.json(result);
+// };
 
-module.exports = {
-    getContactById: ctrlWrapper(getContactById)
-};
+// module.exports = {
+//     getContactById: ctrlWrapper(getContactById)
+// };
