+++ controllers/Contacts/index.js
@@ -1,16 +1,16 @@
-const { listContacts } = require('./listContacts');
-const { getContactById } = require('./getContactById');
-const { addContact } = require('./addContact');
-const { updateContact } = require('./updateContact');
-const { updateFavorite } = require('./updateFavorite');
-const { removeContact } = require('./removeContact');
+// const { listContacts } = require('./listContacts');
+// const { getContactById } = require('./getContactById');
+// const { addContact } = require('./addContact');
+// const { updateContact } = require('./updateContact');
+// const { updateFavorite } = require('./updateFavorite');
+// const { removeContact } = require('./removeContact');
 
 
-module.exports = {
-    listContacts,
-    getContactById,
-    addContact,
-    updateContact,
-    updateFavorite,
-    removeContact
-};
+// module.exports = {
+//     listContacts,
+//     getContactById,
+//     addContact,
+//     updateContact,
+//     updateFavorite,
+//     removeContact
+// };
