+++ controllers/Contacts/listContacts.js
@@ -1,9 +1,9 @@
-const Contact = require('../models/contact');
-const ctrlWrapper = require('../../helpers');
+// const Contact = require('../models/contact');
+// const ctrlWrapper = require('../../helpers');
 
-const listContacts = async (req, res) => {
-  const result = await Contact.find();
-  res.json(result);
-};
+// const listContacts = async (req, res) => {
+//   const result = await Contact.find();
+//   res.json(result);
+// };
 
-module.exports = { listContacts: ctrlWrapper(listContacts) };
+// module.exports = { listContacts: ctrlWrapper(listContacts) };
