+++ controllers/Contacts/removeContact.js
@@ -1,14 +1,14 @@
-const Contact = require("../models/contact");
-const HttpError = require("../helpers/HttpError");
-const ctrlWrapper = require("../helpers/ctrlWrapper");
+// const Contact = require("../models/contact");
+// const HttpError = require("../helpers/HttpError");
+// const ctrlWrapper = require("../helpers/ctrlWrapper");
 
-const removeContact = async (req, res) => {
-  const { contactId } = req.params;
-  const result = await Contact.findByIdAndRemove(contactId);
-  if (!result) {
-    throw HttpError(404, "Not found");
-  }
-  res.status(200).json({ message: "contact deleted" });
-};
+// const removeContact = async (req, res) => {
+//   const { contactId } = req.params;
+//   const result = await Contact.findByIdAndRemove(contactId);
+//   if (!result) {
+//     throw HttpError(404, "Not found");
+//   }
+//   res.status(200).json({ message: "contact deleted" });
+// };
 
-module.exports = {removeContact: ctrlWrapper(removeContact)};
+// module.exports = {removeContact: ctrlWrapper(removeContact)};
