+++ controllers/Contacts/updateContact.js
@@ -1,16 +1,16 @@
-const Contact = require("../models/contact");
-const { ctrlWrapper } = require("./helpers/ctrlWrapper");
-const HttpError = require("../helpers/HttpError");
+// const Contact = require("../models/contact");
+// const { ctrlWrapper } = require("./helpers/ctrlWrapper");
+// const HttpError = require("../helpers/HttpError");
 
-const updateContact = async (req, res) => {
-  const { contactId } = req.params;
-  const result = await Contact.findByIdAndUpdate(contactId, req.body, {
-    new: true,
-  });
-  if (!result) {
-    throw HttpError(404, "Not found");
-  }
-  res.status(200).json(result);
-};
+// const updateContact = async (req, res) => {
+//   const { contactId } = req.params;
+//   const result = await Contact.findByIdAndUpdate(contactId, req.body, {
+//     new: true,
+//   });
+//   if (!result) {
+//     throw HttpError(404, "Not found");
+//   }
+//   res.status(200).json(result);
+// };
 
-module.exports = { updateContact: ctrlWrapper(updateContact) };
+// module.exports = { updateContact: ctrlWrapper(updateContact) };
